+++ CHANGELOG.md
@@ -7,6 +7,11 @@
   * (theimo1221) Add option to run specific ac's regardless of energy level
   * (theimo1221) Add option to disable dachs altogether
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Prevent temperature overcorrection on low outdoor temperatures
+
 ## 3.2.22 (2025-11-30)
 
 * (theimo1221) Prevent dachs start command on high heat Storage